Satis version first try
I added a satis function for retrieving satis information. Currently it is just returning static urls but later I would like satis to be globally available in which case I would change the output to be more dynamic. Also I incorporated the createProject.sh script. Currently only the directory to the script and the target directory for the created project are using variables. The rest of the parameters for createProject.sh are static.